+++ ListSmarter/Services/PersonService.cs
@@ -11,10 +11,12 @@ namespace ListSmarter.Services
     {
         private readonly IPersonRepository _personRepository;
         private readonly IValidator<PersonDto> _personValidator;
+        private readonly ITaskRepository _taskRepository;
 
-        public PersonService(IPersonRepository personRepository,IValidator<PersonDto> personValidator){
+        public PersonService(IPersonRepository personRepository,IValidator<PersonDto> personValidator,ITaskRepository taskRepository){
             _personRepository=personRepository;
             _personValidator=personValidator;
+            _taskRepository=taskRepository;
         }
 
         public List<PersonDto> getAllPeople()
@@ -34,7 +36,13 @@ namespace ListSmarter.Services
 
         public void DeletePerson(int id)
         {
+           List<Models.Task> userTasks = _taskRepository.GetPersonTasks(id);
+           if(userTasks.Count > 0){
+            throw new Exception("Can not delete a person with a task");
+           }else{
             _personRepository.delete(id);
+           }
+            
         }
 
         public void EditPerson(int id, PersonDto data)
